Moved LiterateCoding directives into Salt domain
