+++ doc/conf.py
@@ -130,7 +130,6 @@ exclude_patterns = ['_build', '_incl/*', 'ref/cli/_includes/*.rst']
 
 extensions = [
     'saltdomain', # Must come early
-    'lit_code',
     'sphinx.ext.autodoc',
     'sphinx.ext.autosummary',
     'sphinx.ext.extlinks',
